requirements.txt
jax
jaxlib
dm-haiku
optax
jraph
pyvista
meshio
matplotlib
numpy
